댓글 삭제 만드는중중

+++ final_pjt/back/countries/views.py
@@ -39,8 +39,11 @@ def detail_page(request, country_pk):
 @permission_classes([IsAuthenticated])
 def comment_delete(request, comment_pk):
     comment = get_object_or_404(Comment, pk=comment_pk)
-    comment.delete()
-    return Response(status=status.HTTP_204_NO_CONTENT)
+    if comment.user == request.user:
+        comment.delete()
+        return Response(status=status.HTTP_204_NO_CONTENT)
+    else:
+        return Response({'detail': 'You do not have permission to delete this comment.'}, status=status.HTTP_403_FORBIDDEN)
 
 @api_view(['GET'])
 def profile_page(request, username):
